refactor: move function to helpers
